# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GeoSafe API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/spatial/nearest-depot": {
      "get": {
        "tags": [
          "spatial",
          "spatial"
        ],
        "summary": "Nearest Depot With Required Item",
        "description": "Find the nearest active depot that has the requested item in stock.\nUses PostGIS ST_DWithin and ST_Distance over geography for meter-accurate distance.\nRate limited: 20 requests / 60 seconds per IP.",
        "operationId": "nearest_depot_with_required_item_api_v1_spatial_nearest_depot_get",
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 90,
              "minimum": -90,
              "description": "User latitude",
              "title": "Lat"
            },
            "description": "User latitude"
          },
          {
            "name": "lon",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 180,
              "minimum": -180,
              "description": "User longitude",
              "title": "Lon"
            },
            "description": "User longitude"
          },
          {
            "name": "item_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Required item name",
              "title": "Item Name"
            },
            "description": "Required item name"
          },
          {
            "name": "radius_km",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "exclusiveMinimum": 0,
              "description": "Search radius in kilometers",
              "default": 10.0,
              "title": "Radius Km"
            },
            "description": "Search radius in kilometers"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/spatial/nearest-safe-zone": {
      "get": {
        "tags": [
          "spatial",
          "spatial"
        ],
        "summary": "Nearest Safe Zone",
        "description": "GS-031 — Capacity-aware nearest safe zone lookup.\n\nReturns safe zones ordered by distance. Zones with status='full' are\nincluded but de-prioritised (sorted after non-full zones). Zones with\nstatus='inactive' or 'closed' are excluded entirely.\n\nFalls back to a non-spatial distance calculation when the safe_zone\nlocation column is NULL (e.g. SQLite / missing geometry migrations).",
        "operationId": "nearest_safe_zone_api_v1_spatial_nearest_safe_zone_get",
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 90,
              "minimum": -90,
              "description": "User latitude",
              "title": "Lat"
            },
            "description": "User latitude"
          },
          {
            "name": "lon",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 180,
              "minimum": -180,
              "description": "User longitude",
              "title": "Lon"
            },
            "description": "User longitude"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 20,
              "minimum": 1,
              "description": "Max results",
              "default": 5,
              "title": "Limit"
            },
            "description": "Max results"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/spatial/heatmap": {
      "get": {
        "tags": [
          "spatial",
          "spatial"
        ],
        "summary": "Incident Heatmap",
        "description": "GS-063 — Returns heatmap intensity points as [[lat, lon, weight], ...].\nIncidents are weighted by status; check-ins carry uniform weight 0.5.\nRecords marked spam/dismissed are excluded.",
        "operationId": "incident_heatmap_api_v1_spatial_heatmap_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "incidents | checkins | both",
              "default": "incidents",
              "title": "Source"
            },
            "description": "incidents | checkins | both"
          },
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 365,
              "minimum": 1,
              "description": "Include records from the last N days",
              "default": 30,
              "title": "Days"
            },
            "description": "Include records from the last N days"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/token"
          }
        }
      }
    }
  }
}